refactor(LearningCircle):Added Interface of members
