+++ src/modules/Dashboard/modules/LearningCircle/pages/LearningCircle.tsx
@@ -3,25 +3,9 @@ import styles from "./LearningCircle.module.css";
 import { getLcDetails } from "../services/LearningCircleAPIs";
 import { useParams } from "react-router-dom";
 import pic from "../../Profile/assets/images/dpm.jpg";
+import { LcDetail } from "../services/LearningCircleInterface";
 
-export interface LcMembers {
-    id: string;
-    username: string;
-    profile_pic: string;
-    karma: number;
-}
 
-export interface LcDetail {
-    circle_code: string;
-    college: string;
-    members: LcMembers[];
-    name: string;
-    pending_members: LcMembers[];
-    rank: number;
-    total_karma: number;
-    meet_place: string;
-    meet_time: string;
-}
 
 type Props = {};
