+++ src/modules/Dashboard/modules/LearningCircle/pages/LearningCircleLandingPage.tsx
@@ -6,7 +6,8 @@ import { useNavigate } from "react-router-dom";
 import { useEffect, useState } from "react";
 import { BsChevronRight } from "react-icons/bs";
 import { getUserLearningCircles } from "../services/LearningCircleAPIs";
-import { LcType } from "./LearningCircleFind";
+import { LcType } from "../services/LearningCircleInterface";
+
 
 export const LearningCircleLandingPage = () => {
     const navigate = useNavigate();
